requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
blis==0.7.9
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
chex==0.1.6
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@a9b1bf5920416aaeaec965c25dd9e8f98c864f16
colorama==0.4.6
comm==0.1.2
commonmark==0.9.1
confection==0.0.4
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
debugpy==1.6.6
decorator==4.4.2
defusedxml==0.7.1
dm-tree==0.1.8
easydict==1.10
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.5.0/en_core_web_lg-3.5.0-py3-none-any.whl
executing==1.2.0
fairscale==0.4.13
fastjsonschema==2.16.3
filelock==3.9.0
fire==0.5.0
flatbuffers==1.12
flax==0.5.3
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
ftfy==6.1.1
fvcore==0.1.5.post20221221
gast==0.4.0
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
huggingface-hub==0.12.1
idna==3.4
imageio==2.4.1
imageio-ffmpeg==0.4.5
iopath==0.1.10
ipdb==0.13.11
ipykernel==6.21.2
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jax==0.4.4
jaxlib==0.4.4+cuda11.cudnn82
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.2
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.3.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
langcodes==3.3.0
libclang==15.0.6.1
llama @ file:///home/jeongeun/lm_unct
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
mistune==2.0.5
moviepy==0.2.3.5
msgpack==1.0.4
multidict==6.0.4
murmurhash==1.0.9
nbclassic==0.5.2
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
openai==0.27.0
opencv-python==4.7.0.72
opt-einsum==3.3.0
optax==0.1.4
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathy==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
portalocker==2.7.0
preshed==3.0.8
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybullet==3.2.5
pycparser==2.21
pydantic==1.10.5
Pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==11.2.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.0
sentencepiece==0.1.97
six==1.16.0
sklearn==0.0.post1
slacker==0.14.0
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.4
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
stack-data==0.6.2
style==1.1.0
tabulate==0.9.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.31.0
tensorstore==0.1.33
termcolor==2.2.0
terminado==0.17.1
tf-estimator-nightly==2.8.0.dev2021122109
thinc==8.1.7
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
torchsummary==1.5.1
torchtext==0.14.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
transformers==4.26.1
typer==0.7.0
typing_extensions==4.5.0
update==0.0.1
uri-template==1.2.0
urllib3==1.26.14
wasabi==1.1.1
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.5
wrapt==1.15.0
yacs==0.1.8
yarl==1.8.2
